Upgrade june172021

+++ src/PhpSpreadsheet/Shared/Font.php
@@ -265,10 +265,10 @@ class Font
         }
 
         // Convert from pixel width to column width
-        $columnWidth = Drawing::pixelsToCellDimension($columnWidth, $defaultFont);
+        $columnWidth = Drawing::pixelsToCellDimension((int) $columnWidth, $defaultFont);
 
         // Return
-        return round($columnWidth, 6);
+        return (int) round($columnWidth, 6);
     }
 
     /**
